add header file for GTK window
