+++ main.c
@@ -4,16 +4,7 @@
 #include <gtk/gtk.h>
 #include "input.h"
 #include "sudoku.h"
-
-static void activate(GtkApplication *app, gpointer user_data) {
-	GtkWidget *window;
-	
-	window = gtk_application_window_new(app);
-	gtk_window_set_title(GTK_WINDOW (window), "Sudoku");
-	gtk_window_set_default_size(GTK_WINDOW (window), 200, 200);
-
-	gtk_widget_show_all(window);
-}
+#include "window.h"
 
 int main(int argc, char **argv) {
 
@@ -140,13 +131,7 @@ int main(int argc, char **argv) {
 	// start GTK application if no command-line flags, like input or debug
 	else if (window){
 		puts("Ran GTK app");
-		GtkApplication *app;
-		int status;
-
-		app = gtk_application_new("org.github.jythonscript.sudoku", G_APPLICATION_FLAGS_NONE);
-		g_signal_connect(app, "activate", G_CALLBACK (activate), NULL);
-		status = g_application_run(G_APPLICATION (app), argc, argv);
-		g_object_unref(app);
+		app_new(argc, argv);
 	}
 
 	return 0;
